moved source files under src folder. deleted commented code.
