+++ Tasks/PackerBuild/src/packerTemplateManager.ts
