+++ Tasks/PackerBuild/src/templateFileProviders.ts
