+++ Tasks/PackerBuild/src/templateVariablesProviders.ts
